### 2026-01-04 #739 @thisisamirv
Pre-submission inquiry: rfastlowess
Submitting Author Name: Amir Valizadeh
Submitting Author Github Handle: @thisisamirv
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/thisisamirv/rfastlowess
Submission type: Pre-submission

---

` ``
Package: rfastlowess
Title: 'Rust'-Based High-Performance Parallel LOWESS Smoothing
Version: 0.99.2
Authors@R: person("Amir", "Valizadeh", email = "thisisamirv@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0001-5983-8527"))
Description: Provides high-performance LOWESS smoothing (Locally Weighted Scatterplot Smoothing)
    using a 'Rust' backend. Supports various weight functions, robustness iterations,
    streaming/online processing, cross-validation, and uncertainty quantification.
    Applicable to genomic data, time series, and general-purpose smoothing tasks.
License: MIT + file LICENSE | Apache License (== 2.0)
SystemRequirements: 'Rust' programming language (via 'rustup' or system tools)
biocViews: Preprocessing, StatisticalMethod, Regression, FunctionalPrediction, QualityControl
Suggests: BiocStyle, knitr, rmarkdown, testthat (>= 3.0.0)
` ``

The package provides a high-performance Rust wrapper for LOWESS (Locally Weighted Scatterplot Smoothing), which is a fundamental regression and EDA technique. It extends the standard implementation with parallel execution, robust statistics, and uncertainty quantification.

No (not yet implemented, happy to add if required for this category).

The target audience includes researchers and data scientists working with large datasets, such as in genomics (ChIP-seq/ATAC-seq normalization) or high-frequency time series, where standard R implementations are too slow. It offers significant speedups (2-6x) via parallel processing.

Yes, `stats::lowess` provides the standard implementation. `rfastlowess` differs by offering:
1.  **Parallelism**: Uses Rayon for multi-threaded execution.
2.  **Robustness options**: Supports MAD-based scale estimation and different weight functions (Bisquare, Huber, etc.).
3.  **Uncertainty Quantification**: Calculates confidence and prediction intervals (missing in base R).
4.  **Streaming/Online modes**: Handles data larger than memory or real-time streams.

Yes (N/A, pure numerical software).

The package requires a Rust toolchain to compile from source, but we provide binary builds via R-Universe to ease installation.
