### 2026-01-12 #739 @mpadge
@thisisamirv This looks like a very exciting submission, as it would extend our review to consider R packages as bindings to larger suites of implementations in other languages. It definitely looks in scope for statistical software in the *Regression and Supervised Learning* category, and possibly also the *EDA* category, as you've indicated. One concern I would have is that a review would need to focus almost entirely on the Rust code, rather than the very thin R bindings. I'll discuss with the editorial team whether we think this might stray too far outside our core R expertise. A further issue at present would be that our automated compliance assessment with our [statistical standards](https://stats-devguide.ropensci.org) assumes compliance statements from code in R, C, or C++ only. We'd need to update that to include Rust code too. That shouldn't be too difficult, and would be a useful extension anyway, but we'll have to wait to first see what the editorial teams thinks. I'll get back to you soon.
