### 2026-01-13 #739 @mpadge
@thisisamirv Opinions from editors are unanimously :+1: in favour of proceeding here. We'll first need to update a few aspects of our checking system to get it to automatically work on your package without needing [your current hard-coded paths](https://github.com/thisisamirv/lowess-project/blob/aca86b2fea6bef50750ba90dc741d16cd36404fa/.github/workflows/ci.yml#L179). That's the easy bit...

A more difficult aspect is going to be automatically parsing your code, which appears to not be currently possible, as far as I can tell. The source code is [extracted from a tarball](https://github.com/thisisamirv/lowess-project/blob/aca86b2fea6bef50750ba90dc741d16cd36404fa/bindings/r/src/Makevars.in#L18). While it would be possible to parse your `configure` and `Makevars.in` files to locate source files in the repo, they are ultimately contained within a binary object. That can't be parsed, and so our checking systems wouldn't be able to proceed from there.

This is the first time I've had to consider the implications of complex package structures for our automated checking systems, so nothing I say here should be considered definitive or irrevocable. That said, it seems to me that progress would only be possible if you could keep *most* of that configure structure intact, yet get the contents of `vendor/lowess` and `vendor/fastLowess` directly from the corresponding code in `./crates`, along the lines of your fallback code in Julia. Do you think that would be possible? (In current form, having duplicated code in `crates` and then dumped in `bindings/r/src/vendor.tar.xz` is risky anyway, even with your exemplary Makefile to control it.)

And if so, while you're doing it, could you please think about the best way for us to automatically parse where that code is coming from. Again, the Julia fallback code is a good example that would be possible to follow, but even that is only run after the crate has been built, so wouldn't provide a direct path to the actual source code.

---

In short: I do hope that we'll be able to proceed here, but will definitely need your help in figuring out how 😄  Thanks in advance :+1:
