### 2026-02-01 #739 @thisisamirv
@mpadge  @JosiahParry  Thank you both for the excellent feedback. I have implemented the suggested changes:

1. SOURCE_MANIFEST.toml: I've added a TOML manifest in src/ that maps the vendored crates to their source locations in the repository root. Following Josiah's advice, I've also added this to **.Rbuildignore** to keep it out of the CRAN-destined tarball while remaining available for your GitHub-based automated checks.

2. Portability Fixes: I've refactored  **Makevars.in** and **Makevars.win** to use portable shell commands (e.g., using $(TAR) -xf and limiting cargo build -j 2). This should address the CRAN compatibility concerns.

3. srrstats tags: I have no impediments to adding @srrstats tags to the Rust source code. I have already started by adding initial tags to the R package documentation (**rfastlowess-package.R**) and the core Rust-to-R boundary code (**lib.rs**). Using inner comments //! for module-level compliance statements and non-indented /// for others sounds like a very clean and parseable approach.

Please let me know if there are any other technical updates you'd like to see as you refine the srr parsing system for Rust!
