### 2026-02-04 #739 @mpadge
@thisisamirv Parsing of `srr` standards statements in Rust is definitely going to be viable. I've already got the main mechanism in, and all works. So you may definitely start adding compliance statements to your main rust src code, in one of these forms (outer and inner comments):
``` rust
\\\ @srrstats {G1.1} Compliance statements ...
\\! @srrstatsNA {G1.2} Not applicable here  ...
```
Our system won't get read your `SOURCE_MANIFEST` file, but the inner mechanics are there, and that will be just a minor tweak that can be implemented once you've got that in place.

So it's back over to you, and I look forward to hearing progress. Let us know of any questions in the meantime.
