### 2026-03-13 #739 @mpadge
@thisisamirv I'm going to re-open this to wait for a response from @emilyriederer. There may be outstanding issues from the EiC side, so best to clarify potential final steps here before we move on. Thanks 👍
